rename ggml-cpu-quants.c
