+++ ggml/src/ggml-cpu/CMakeLists.txt
@@ -14,8 +14,8 @@ function(ggml_add_cpu_backend_variant_impl tag_name)
         ggml-cpu/repack.h
         ggml-cpu/ggml-cpu-hbm.cpp
         ggml-cpu/ggml-cpu-hbm.h
-        ggml-cpu/ggml-cpu-quants.c
-        ggml-cpu/ggml-cpu-quants.h
+        ggml-cpu/quants.c
+        ggml-cpu/quants.h
         ggml-cpu/ggml-cpu-traits.cpp
         ggml-cpu/ggml-cpu-traits.h
         ggml-cpu/amx/amx.cpp
